backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
blinker==1.9.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
fastapi==0.135.1
Flask==3.1.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2-binary==2.9.9
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
sqlmodel==0.0.16
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn[standard]==0.29.0
Werkzeug==3.1.6
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic-settings>=2.0.0
pytest>=8.0.0
httpx>=0.27.0